refactor(core/reindent): move reindent to utils.
Also improve docs and TS
