+++ src/core/reindent.js
@@ -1,28 +1,12 @@
 // @ts-check
 /**
- * Module core/reindent
+ * @module core/reindent
  *
- * Removes common indents across the IDL texts,
+ * Normalizes indents across the pre elements in the document,
  * so that indentation inside <pre> won't affect the rendered result.
  */
-
+import { reindent } from "./utils.js";
 export const name = "core/reindent";
-
-/**
- * @param {string} text
- */
-export function reindent(text) {
-  if (!text) {
-    return text;
-  }
-  const lines = text.trimEnd().split("\n");
-  while (lines.length && !lines[0].trim()) {
-    lines.shift();
-  }
-  const indents = lines.filter(s => s.trim()).map(s => s.search(/[^\s]/));
-  const leastIndent = Math.min(...indents);
-  return lines.map(s => s.slice(leastIndent)).join("\n");
-}
 
 export function run() {
   for (const pre of document.getElementsByTagName("pre")) {
